gpt-tts/pyproject.toml
[project]
name = "gpt-tts"
version = "0.1.0"
description = ""
authors = [
    {name = "irayspace",email = "irayspacii@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=1.97.1,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.47.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
